fix: do not surface parse error parsing certain jsx in .js files
Closes #373
